backend/requirements.txt
# === Phase 1: Core ===
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
celery[redis]==5.3.6
redis==5.0.1
minio==7.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
python-multipart==0.0.9
psycopg2-binary==2.9.9

# === Phase 2: Ingestion ===
# Data processing
pandas==2.2.0
pyarrow==15.0.0
openpyxl==3.1.2

# Document parsing
pdfplumber==0.11.0
pytesseract==0.3.10
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# File detection & encoding
python-magic==0.4.27
chardet==5.2.0

# Text analysis
langdetect==1.0.9
tiktoken==0.7.0

# Connectors
huggingface-hub==0.23.0
datasets==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
boto3==1.34.49

# Async utilities
aiofiles==23.2.1
websockets==12.0
